# Framework throughput improvements