--- a/examples/multibody/cassie_benchmark/record_results.sh
+++ b/examples/multibody/cassie_benchmark/record_results.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-# Collect context information for a benchmark experiment.
-
-${TEST_SRCDIR}/drake/tools/performance/record_results \
-    ${TEST_SRCDIR}/drake/examples/multibody/cassie_benchmark/cassie_bench \
-    "$@"