This is a WIP branch that will be continuously refined and flow into PRs.

The overall effect is a >2x throughput speed-up with vector PassThrough systems in a small diagram, i.e., cutting the framework overhead in half.

- [x] #15418
- [x] #15419
- [x] #15420
- [x] #15422
- [x] #15431
- [x] #15432
- [x] #15470

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/15421)
<!-- Reviewable:end -->