@jwnimmer-tri 2021-07-27
All of the performance-related patches have been merged.  (For the benchmarking tool improvements, I'll open a separate PR shortly.)  Here are the final results from these changes (using the as-of-yet unmerged benchmarking tool, currently in draft form in this PR).

Current master (as of a7d330ee0076ccdd42aff5576676e9dfcc2423bf):

```
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
BasicFixture/PassThrough3_mean          492 ns          492 ns            9
BasicFixture/PassThrough3_median        495 ns          495 ns            9
BasicFixture/PassThrough3_stddev       6.20 ns         6.20 ns            9
BasicFixture/PassThrough3_min           484 ns          484 ns            9
BasicFixture/PassThrough3_max           501 ns          501 ns            9
```

After reverting the improvements:
- git revert d0e6a820ff01e9531ff3ef321a1a5d42d8b60930
- git revert 4e6ef08bfe5778127c700adb56ae17d871507819
- git revert d52d1d216398ffc0966384995965a378926ee1fa
- git revert bb004802b6aed266916a1a5aa28db034bdfb4b97
- git revert 28a086813c722972df41a30c37617e118ece09a7

```
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
BasicFixture/PassThrough3_mean         1125 ns         1125 ns            9
BasicFixture/PassThrough3_median       1126 ns         1126 ns            9
BasicFixture/PassThrough3_stddev       15.0 ns         15.0 ns            9
BasicFixture/PassThrough3_min          1108 ns         1108 ns            9
BasicFixture/PassThrough3_max          1155 ns         1155 ns            9
```

Taking the `min` timing value, that's down from 1108ns to 484ns, i.e., a 56% reduction in time, or +129% improvement in throughput.